payload-receiver-service/tests/integration/requirements.txt
pytest==3.8.1
requests==2.22.0
asyncio-nats-client===0.11.4
opni-nats==0.0.0.4
faker
